exit early if layers WASM layers not found
