+++ crates/containerd-shim-wasm/src/sandbox/containerd/mod.rs
@@ -1,0 +1,4 @@
+mod client;
+mod lease;
+
+pub(crate) use client::Client;
